Finalizing last conformance checks: hmpg compatibility checks (#162) and Fixing Date checkin Issue (#167)

See merge request MPEG/3dgh/haptics/software/ReferenceSoftwarePhase1!147

+++ source/IOHaptics/include/IOCompatibility.h
@@ -350,7 +350,7 @@ public:
   };
 
   static auto checkSync(types::Sync &sync, std::vector<std::string> &logs) -> void {
-    if (static_cast<unsigned int>(sync.getTimestamp()) > MAX_32_BITS_UNSIGNED) {
+    if (sync.getTimestamp() > MAX_32_BITS_UNSIGNED) {
       logs.push_back(hjifWarningCodeToString.at(hjifWarningCode::Sync_Timestamp_OutOfRange));
     }
