+++ source/Types/include/Channel.h
@@ -111,10 +111,10 @@ public:
   auto Evaluate(double position, unsigned int timescale) -> double;
   auto EvaluateChannel(uint32_t sampleCount, int fs, int pad, unsigned int timescale)
       -> std::vector<double>;
-  [[nodiscard]] auto getFrequencySampling() const -> std::optional<uint32_t>;
-  auto setFrequencySampling(std::optional<uint32_t> newFrequencySampling) -> void;
-  [[nodiscard]] auto getSampleCount() const -> std::optional<uint32_t>;
-  auto setSampleCount(std::optional<uint32_t> newSampleCount) -> void;
+  [[nodiscard]] auto getFrequencySampling() const -> std::optional<uint64_t>;
+  auto setFrequencySampling(std::optional<uint64_t> newFrequencySampling) -> void;
+  [[nodiscard]] auto getSampleCount() const -> std::optional<uint64_t>;
+  auto setSampleCount(std::optional<uint64_t> newSampleCount) -> void;
   [[nodiscard]] auto getDirection() const -> std::optional<Vector>;
   auto setDirection(std::optional<Vector> newDirection) -> void;
   [[nodiscard]] auto getActuatorResolution() const -> std::optional<Vector>;
@@ -135,8 +135,8 @@ private:
   std::vector<Band> bands = {};
   std::optional<int> priority;
   std::optional<int> referenceDeviceId;
-  std::optional<uint32_t> frequencySampling = std::nullopt;
-  std::optional<uint32_t> sampleCount = std::nullopt;
+  std::optional<uint64_t> frequencySampling = std::nullopt;
+  std::optional<uint64_t> sampleCount = std::nullopt;
   std::optional<Vector> direction = std::nullopt;
   std::optional<Vector> actuatorResolution = std::nullopt;
   std::optional<std::vector<BodyPartTarget>> bodyPartTarget = std::nullopt;
